// packages/controller/src/types.d.ts
export type LogLevel = 'error' | 'warn' | 'info' | 'log';
export type ConsoleEntryType = 'uncaught_error' | 'unhandled_rejection' | 'console_error' | 'console_warn' | 'console_info' | 'console_log';
export interface StackFrame {
    functionName?: string;
    filename?: string;
    lineno?: number;
    colno?: number;
    raw?: string;
}
export interface ConsoleEntry {
    id: string;
    type: ConsoleEntryType;
    level: LogLevel;
    timestamp: number;
    message: string;
    args?: any[];
    stack?: string;
    parsedStack?: StackFrame[];
    count: number;
    firstSeen: number;
    lastSeen: number;
}